requirements.txt.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
bcrypt==4.0.1
PyJWT==2.8.0